removing "external-lights" package
