--- libs/externallights/README.md
@@ -1,3 +1,0 @@
-# externallights
-
-External programmable lights package.
