--- libs/externallights/_locales/advancedlight-jsdoc-strings.json
@@ -1,1 +1,0 @@
-{}
