--- libs/externallights/_locales/advancedlight-strings.json
@@ -1,1 +1,0 @@
-{}
