--- libs/externallights/_locales/externallights-jsdoc-strings.json
@@ -1,27 +1,0 @@
-{
-  "light": "Functions to operate colored LEDs.\n\nFunctions to operate colored LEDs.",
-  "light.NeoPixelStrip": "A NeoPixel strip",
-  "light.NeoPixelStrip.brightness": "Get the brightness of the pixel strip.",
-  "light.NeoPixelStrip.clear": "Turn off all pixel LEDs.",
-  "light.NeoPixelStrip.graph": "Display a vertical bar graph based on the `value` and `high` value.\nIf `high` is 0, the chart gets adjusted automatically.",
-  "light.NeoPixelStrip.length": "Get the number of pixels on the strip",
-  "light.NeoPixelStrip.move": "Shift LEDs forward and clear with zeros.\nYou need to call ``show`` to make the changes visible.",
-  "light.NeoPixelStrip.photonFlip": "Switch the direction of the photon pulse.",
-  "light.NeoPixelStrip.photonForward": "Move a photon effect along the pixel strip by a number of steps.",
-  "light.NeoPixelStrip.pixelColor": "Gets the pixel color.",
-  "light.NeoPixelStrip.range": "Create a range of pixels.",
-  "light.NeoPixelStrip.setAll": "Set all of the pixels on the strip to one RGB color.",
-  "light.NeoPixelStrip.setBrightness": "Set the brightness of the strip. This flag only applies to future operation.",
-  "light.NeoPixelStrip.setBuffered": "Enables or disables automatically calling show when a change is made",
-  "light.NeoPixelStrip.setPhotonColor": "Set the photon color.",
-  "light.NeoPixelStrip.setPhotonMode": "Set the photon mode to pen up, pen down, or eraser.",
-  "light.NeoPixelStrip.setPixelColor": "Set the pixel to a given color.\nYou need to call ``show`` to make the changes visible.",
-  "light.NeoPixelStrip.setPixelWhiteLED": "Set the white brightness of a pixel in a NeoPixel strip of RGB+W LEDs.\nThis only works for RGB+W NeoPixels.",
-  "light.NeoPixelStrip.show": "Make the strip show all the new changes for the pixels.",
-  "light.NeoPixelStrip.showAnimation": "Show an animation or queue an animation in the animation queue",
-  "light.NeoPixelStrip.showAnimationFrame": "Show a single animation frame",
-  "light.NeoPixelStrip.stopAllAnimations": "Stop the current animation and any other animations ready to show.",
-  "light.createStrip": "Create a new NeoPixel driver for `numleds` LEDs.",
-  "light.createStrip|param|numleds": "number of leds in the strip, eg: 24,30,60,64",
-  "light.createStrip|param|pin": "the pin where the neopixel is connected."
-}
