--- libs/externallights/_locales/externallights-strings.json
@@ -1,27 +1,0 @@
-{
-  "light.NeoPixelStrip.brightness|block": "%strip|brightness",
-  "light.NeoPixelStrip.clear|block": "%strip|clear",
-  "light.NeoPixelStrip.graph|block": "%strip|graph of %value |up to %high",
-  "light.NeoPixelStrip.length|block": "%strip|length",
-  "light.NeoPixelStrip.move|block": "%strip|%kind=MoveKind|by %offset",
-  "light.NeoPixelStrip.photonFlip|block": "%strip|photon flip",
-  "light.NeoPixelStrip.photonForward|block": "%strip|photon forward by %steps",
-  "light.NeoPixelStrip.pixelColor|block": "%strip|pixel color at %pixeloffset",
-  "light.NeoPixelStrip.range|block": "%strip|range from %start|with %length|pixels",
-  "light.NeoPixelStrip.setAll|block": "%strip|set all pixels to %rgb=colorNumberPicker",
-  "light.NeoPixelStrip.setBrightness|block": "%strip|set brightness %brightness",
-  "light.NeoPixelStrip.setBuffered|block": "%strip|set buffered  %on",
-  "light.NeoPixelStrip.setPhotonColor|block": "%strip=variables_get| photon set pen color %color=colorWheelPicker",
-  "light.NeoPixelStrip.setPhotonMode|block": "%strip|photon %mode",
-  "light.NeoPixelStrip.setPixelColor|block": "%strip|set pixel color at %pixeloffset|to %rgb=colorNumberPicker",
-  "light.NeoPixelStrip.setPixelWhiteLED|block": "%strip|set pixel white LED at %pixeloffset|to %white",
-  "light.NeoPixelStrip.showAnimationFrame|block": "%strip|show animation frame %animation=light_animation",
-  "light.NeoPixelStrip.showAnimation|block": "%strip|show %animation=light_animation|animation for %duration=timePicker|ms",
-  "light.NeoPixelStrip.show|block": "%strip|show",
-  "light.NeoPixelStrip.stopAllAnimations|block": "%strip|stop all animations",
-  "light.createStrip|block": "create strip|on %pin|with %numleds|pixels",
-  "light|block": "light",
-  "{id:category}Light": "Light",
-  "{id:group}More": "More",
-  "{id:group}Photon": "Photon"
-}
