--- libs/externallights/pxt.json
@@ -1,17 +1,0 @@
-{
-    "name": "externallights",
-    "description": "External programmable lights package",
-    "files": [
-        "README.md",
-        "neopixeloverrides.ts",
-        "create.ts"
-    ],
-    "testFiles": [
-        "test.ts"
-    ],
-    "public": true,
-    "dependencies": {
-        "core": "file:../core",
-        "light": "file:../light"
-    }
-}
