+++ libs/light/_locales/light-jsdoc-strings.json
@@ -38,6 +38,10 @@
   "light.createNeoPixelStrip": "Create a new NeoPixel driver for `numleds` LEDs.",
   "light.createNeoPixelStrip|param|numleds": "number of leds in the strip, eg: 24,30,60,64",
   "light.createNeoPixelStrip|param|pin": "the pin where the neopixel is connected.",
+  "light.createStrip": "Create a new NeoPixel driver for `numleds` LEDs.",
+  "light.createStrip|param|mode": "the light encoding mode for different LED strips, eg: NeoPixelMode.RGB",
+  "light.createStrip|param|numleds": "number of leds in the strip, eg: 24,30,60,64",
+  "light.createStrip|param|pin": "the pin where the neopixel is connected.",
   "light.defaultPin": "Gets the default pin for built in neopixels",
   "light.fade": "Fade the color by the brightness",
   "light.fade|param|brightness": "the amount of brightness to apply to the color, eg: 128",
