+++ libs/light/_locales/light-strings.json
@@ -49,6 +49,7 @@
   "light.colorWheelPicker|block": "%color",
   "light.colors|block": "%color",
   "light.createNeoPixelStrip|block": "create strip|pin %pin|pixels %numleds|mode %mode",
+  "light.createStrip|block": "create strip|on %pin|with %numleds|pixels using %mode",
   "light.fade|block": "fade %color=neopixel_colors|by %brightness",
   "light.hsv|block": "hue %hue|sat %sat|val %val",
   "light.rgb|block": "red %red|green %green|blue %blue",
