+++ pxtarget.json
@@ -20,8 +20,7 @@
         "libs/switch",
         "libs/cable",
         "libs/infrared",
-        "libs/circuit-playground",
-        "libs/externallights"
+        "libs/circuit-playground"
     ],
     "serial": {
         "useHF2": true,
